[aerial_robot_dynamics] separate base class and test class and refactor all code
